Moved helper class inside cpp.
