+++ brayns/utils/image/ImageCodecRegistry.h
@@ -21,6 +21,7 @@
 
 #pragma once
 
+#include <stdexcept>
 #include <string>
 
 #include "ImageCodec.h"
